@@ -3,10 +3,10 @@ description: A new Flutter project.
 
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-version: 1.0.1+3
+version: 1.0.2+4
 
 environment:
-  sdk: ">=2.18.5 <3.0.0"
+   sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
@@ -14,44 +14,40 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.5
+  cupertino_icons: ^1.0.8
   http: ^0.13.5
-  shared_preferences: ^2.0.15
+  shared_preferences: ^2.2.3
   flutter_bloc: ^8.1.1
-  firebase_auth: ^4.2.1
-  firebase_core: ^2.4.0
-  cloud_firestore: ^4.2.0
-  firebase_analytics: ^10.0.7
+  firebase_auth: ^4.20.0
+  firebase_core: ^2.32.0
+  cloud_firestore: ^4.17.5
   equatable: ^2.0.5
-  flutter_screenutil: ^5.6.0
-  sqflite: ^2.2.2
-  path_provider: ^2.0.11
-  path: ^1.8.2
-  share_plus: ^6.3.0
+  path_provider: ^2.1.0
+  path: ^1.9.0
+  share_plus: ^7.2.2
   custom_widgets:
     git:
       url: https://github.com/rojanshr1996/custom_widgets.git
       ref: main
-  provider: ^6.0.5
-  firebase_storage: ^11.0.8
+  provider: ^6.1.2
+  firebase_storage: ^11.7.7
   cached_network_image: ^3.2.3
-  file_picker: ^5.2.4
-  intl: ^0.17.0
-  image_picker: ^0.8.6
-  pdf_render: ^1.3.6
-  google_sign_in: ^5.4.2
-  font_awesome_flutter: ^10.3.0
+  file_picker: ^8.0.6
+  intl: ^0.19.0
+  image_picker: ^0.8.9
+  pdf_render: ^1.4.10
+  google_sign_in: ^5.4.4
+  font_awesome_flutter: ^10.7.0
   like_button: ^2.0.5
-  url_launcher: ^6.1.6
-  flutter_email_sender: ^5.1.0
-  firebase_messaging: ^14.1.4
-  flutter_local_notifications: ^12.0.4
-  flutter_datetime_picker: ^1.5.1
+  url_launcher: ^6.2.6
+  flutter_email_sender: ^5.2.0
+  firebase_messaging: ^14.9.4
+  flutter_local_notifications: ^16.3.3
 
-flutter_icons:
+flutter_launcher_icons:
   android: "launcher_icon"
   ios: true
-  image_path: "assets/splash.png"
+  image_path: "assets/notesterLogo.png"
   remove_alpha_ios: true
 
 
@@ -60,8 +56,8 @@ dev_dependencies:
     sdk: flutter
 
   flutter_lints: ^2.0.1
-  flutter_launcher_icons: ^0.11.0
-  bloc_test: ^9.1.0
+  flutter_launcher_icons: ^0.13.1
+  bloc_test: ^9.1.3
   test: ^1.20.0
 
 flutter:
