@@ -2,6 +2,10 @@ import 'dart:io';
 
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:custom_widgets/custom_widgets.dart';
+import 'package:firebase_storage/firebase_storage.dart';
+import 'package:flutter/material.dart';
+import 'package:intl/intl.dart';
+import 'package:like_button/like_button.dart';
 import 'package:notester/model/model.dart';
 import 'package:notester/services/auth_services.dart';
 import 'package:notester/services/cloud/cloud_note.dart';
@@ -12,11 +16,6 @@ import 'package:notester/utils/custom_text_style.dart';
 import 'package:notester/utils/utils.dart';
 import 'package:notester/view/notes/color_slider.dart';
 import 'package:notester/view/route/routes.dart';
-import 'package:firebase_storage/firebase_storage.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_datetime_picker/flutter_datetime_picker.dart';
-import 'package:intl/intl.dart';
-import 'package:like_button/like_button.dart';
 import 'package:path/path.dart' as path;
 import 'package:path/path.dart';
 
@@ -74,7 +73,8 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
     final title = _titleController.text.trim();
     await _notesService.updateNote(
       documentId: note.documentId,
-      createdDate: _createdDate.value == "" ? "${DateTime.now()}" : _createdDate.value,
+      createdDate:
+          _createdDate.value == "" ? "${DateTime.now()}" : _createdDate.value,
       // createdDate: "${DateTime.now()}",
       text: text,
       title: title,
@@ -95,7 +95,8 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
     final title = _titleController.text.trim();
     await _notesService.updateNote(
       documentId: note.documentId,
-      createdDate: _createdDate.value == "" ? "${DateTime.now()}" : _createdDate.value,
+      createdDate:
+          _createdDate.value == "" ? "${DateTime.now()}" : _createdDate.value,
       // createdDate: "${DateTime.now()}",
       text: text,
       title: title,
@@ -168,7 +169,10 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
 
   void _deleteNoteIfTextIsEmpty() {
     final note = _note;
-    if (_textController.text.isEmpty && note != null && _imageUrl.value == "" && _fileUrl.value == "") {
+    if (_textController.text.isEmpty &&
+        note != null &&
+        _imageUrl.value == "" &&
+        _fileUrl.value == "") {
       _notesService.deleteNote(documentId: note.documentId);
     }
   }
@@ -182,7 +186,8 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
     if (note != null && text.isNotEmpty) {
       await _notesService.updateNote(
         documentId: note.documentId,
-        createdDate: _createdDate.value == "" ? "${DateTime.now()}" : _createdDate.value,
+        createdDate:
+            _createdDate.value == "" ? "${DateTime.now()}" : _createdDate.value,
         // createdDate: "${DateTime.now()}",
         text: text,
         title: title,
@@ -259,12 +264,15 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                   Utilities.closeActivity(context);
                 },
                 style: ButtonStyle(
-                  backgroundColor: MaterialStateProperty.all<Color>(Theme.of(context).scaffoldBackgroundColor),
+                  backgroundColor: WidgetStateProperty.all<Color>(
+                      Theme.of(context).scaffoldBackgroundColor),
                 ),
                 child: Text(
                   "Save",
-                  style:
-                      Theme.of(context).textTheme.bodyMedium!.copyWith(color: Theme.of(context).colorScheme.background),
+                  style: Theme.of(context)
+                      .textTheme
+                      .bodyMedium!
+                      .copyWith(color: Theme.of(context).colorScheme.surface),
                 ),
               ),
             ),
@@ -285,7 +293,8 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                           mainAxisSize: MainAxisSize.min,
                           children: [
                             Padding(
-                              padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 10),
+                              padding: const EdgeInsets.symmetric(
+                                  horizontal: 16, vertical: 10),
                               child: Row(
                                 children: [
                                   Expanded(
@@ -293,7 +302,9 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                                       icon: Icon(
                                         Icons.image,
                                         size: 18,
-                                        color: Theme.of(context).colorScheme.background,
+                                        color: Theme.of(context)
+                                            .colorScheme
+                                            .surface,
                                       ),
                                       onPressed: () {
                                         FocusScope.of(context).unfocus();
@@ -301,8 +312,14 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                                       },
                                       label: Text(
                                         "Add Image",
-                                        style: Theme.of(context).textTheme.bodyMedium!.copyWith(
-                                            color: Theme.of(context).colorScheme.background, fontWeight: medium),
+                                        style: Theme.of(context)
+                                            .textTheme
+                                            .bodyMedium!
+                                            .copyWith(
+                                                color: Theme.of(context)
+                                                    .colorScheme
+                                                    .surface,
+                                                fontWeight: medium),
                                       ),
                                     ),
                                   ),
@@ -312,7 +329,9 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                                       icon: Icon(
                                         Icons.file_copy,
                                         size: 18,
-                                        color: Theme.of(context).colorScheme.background,
+                                        color: Theme.of(context)
+                                            .colorScheme
+                                            .surface,
                                       ),
                                       onPressed: () {
                                         FocusScope.of(context).unfocus();
@@ -320,8 +339,14 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                                       },
                                       label: Text(
                                         "Add File",
-                                        style: Theme.of(context).textTheme.bodyMedium!.copyWith(
-                                            color: Theme.of(context).colorScheme.background, fontWeight: medium),
+                                        style: Theme.of(context)
+                                            .textTheme
+                                            .bodyMedium!
+                                            .copyWith(
+                                                color: Theme.of(context)
+                                                    .colorScheme
+                                                    .surface,
+                                                fontWeight: medium),
                                       ),
                                     ),
                                   ),
@@ -334,7 +359,8 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                                   crossAxisAlignment: CrossAxisAlignment.start,
                                   children: [
                                     Padding(
-                                      padding: const EdgeInsets.symmetric(horizontal: 16),
+                                      padding: const EdgeInsets.symmetric(
+                                          horizontal: 16),
                                       child: Row(
                                         children: [
                                           ValueListenableBuilder(
@@ -344,65 +370,88 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                                                 child: _createdDate.value == ""
                                                     ? const SizedBox()
                                                     : Text(
-                                                        DateFormat.yMMMMd().format(DateTime.parse(_createdDate.value)),
-                                                        style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                                                            fontWeight: medium, color: Theme.of(context).hintColor),
+                                                        DateFormat.yMMMMd()
+                                                            .format(
+                                                                DateTime.parse(
+                                                                    _createdDate
+                                                                        .value)),
+                                                        style: Theme.of(context)
+                                                            .textTheme
+                                                            .bodySmall
+                                                            ?.copyWith(
+                                                                fontWeight:
+                                                                    medium,
+                                                                color: Theme.of(
+                                                                        context)
+                                                                    .hintColor),
                                                       ),
                                               );
                                             },
                                           ),
                                           Column(
-                                            crossAxisAlignment: CrossAxisAlignment.end,
+                                            crossAxisAlignment:
+                                                CrossAxisAlignment.end,
                                             children: [
-                                              OutlinedButton.icon(
-                                                icon: Icon(
-                                                  Icons.access_time,
-                                                  size: 16,
-                                                  color: Theme.of(context).colorScheme.background,
-                                                ),
-                                                onPressed: () {
-                                                  DatePicker.showDateTimePicker(
-                                                    context,
-                                                    showTitleActions: true,
-                                                    minTime: DateTime.now(),
-                                                    currentTime: DateTime.now(),
-                                                    maxTime: DateTime(2200, 1, 1),
-                                                    onConfirm: (date) {
-                                                      _reminder.value = date.toString();
-                                                    },
-                                                    locale: LocaleType.en,
-                                                  );
-                                                },
-                                                label: Text(
-                                                  "Reminder",
-                                                  style: Theme.of(context).textTheme.bodySmall!.copyWith(
-                                                      color: Theme.of(context).colorScheme.background,
-                                                      fontWeight: medium),
-                                                ),
-                                              ),
+                                              // OutlinedButton.icon(
+                                              //   icon: Icon(
+                                              //     Icons.access_time,
+                                              //     size: 16,
+                                              //     color: Theme.of(context).colorScheme.surface,
+                                              //   ),
+                                              //   onPressed: () {
+                                              //     date_time_picker.DatePicker.showDateTimePicker(
+                                              //       context,
+                                              //       showTitleActions: true,
+                                              //       minTime: DateTime.now(),
+                                              //       currentTime: DateTime.now(),
+                                              //       maxTime: DateTime(2200, 1, 1),
+                                              //       onConfirm: (date) {
+                                              //         _reminder.value = date.toString();
+                                              //       },
+                                              //       locale: date_time_picker.LocaleType.en,
+                                              //     );
+                                              //   },
+                                              //   label: Text(
+                                              //     "Reminder",
+                                              //     style: Theme.of(context).textTheme.bodySmall!.copyWith(
+                                              //         color: Theme.of(context).colorScheme.surface, fontWeight: medium),
+                                              //   ),
+                                              // ),
                                               ValueListenableBuilder(
                                                 valueListenable: _reminder,
-                                                builder: (context, reminder, _) {
+                                                builder:
+                                                    (context, reminder, _) {
                                                   return Row(
                                                     children: [
                                                       _reminder.value == ""
                                                           ? const SizedBox()
                                                           : InkWell(
                                                               onTap: () {
-                                                                _reminder.value = "";
+                                                                _reminder
+                                                                    .value = "";
                                                               },
-                                                              child: const Icon(Icons.close,
-                                                                  size: 14, color: AppColors.cRed),
+                                                              child: const Icon(
+                                                                  Icons.close,
+                                                                  size: 14,
+                                                                  color:
+                                                                      AppColors
+                                                                          .cRed),
                                                             ),
                                                       const SizedBox(width: 8),
                                                       Text(
                                                         _reminder.value == ""
                                                             ? ""
                                                             : "${DateFormat.yMMMMd().format(DateTime.parse(_reminder.value))}, ${DateFormat.jm().format(DateTime.parse(_reminder.value))}",
-                                                        style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                                                            fontWeight: medium,
-                                                            fontSize: 10,
-                                                            color: Theme.of(context).hintColor),
+                                                        style: Theme.of(context)
+                                                            .textTheme
+                                                            .bodySmall
+                                                            ?.copyWith(
+                                                                fontWeight:
+                                                                    medium,
+                                                                fontSize: 10,
+                                                                color: Theme.of(
+                                                                        context)
+                                                                    .hintColor),
                                                       ),
                                                     ],
                                                   );
@@ -414,17 +463,20 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                                       ),
                                     ),
                                     Padding(
-                                      padding: const EdgeInsets.only(left: 16, right: 12),
+                                      padding: const EdgeInsets.only(
+                                          left: 16, right: 12),
                                       child: Row(
                                         children: [
                                           Padding(
-                                            padding: const EdgeInsets.only(right: 8),
+                                            padding:
+                                                const EdgeInsets.only(right: 8),
                                             child: Container(
                                               height: 40,
                                               width: 40,
                                               decoration: BoxDecoration(
                                                   shape: BoxShape.circle,
-                                                  border: Border.all(color: AppColors.cWhite),
+                                                  border: Border.all(
+                                                      color: AppColors.cWhite),
                                                   color: _color.value),
                                             ),
                                           ),
@@ -432,17 +484,29 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                                           Expanded(
                                             child: TextField(
                                               controller: _titleController,
-                                              keyboardType: TextInputType.multiline,
-                                              style:
-                                                  Theme.of(context).textTheme.titleMedium?.copyWith(fontWeight: bold),
+                                              keyboardType:
+                                                  TextInputType.multiline,
+                                              style: Theme.of(context)
+                                                  .textTheme
+                                                  .titleMedium
+                                                  ?.copyWith(fontWeight: bold),
                                               maxLines: null,
-                                              textCapitalization: TextCapitalization.sentences,
+                                              textCapitalization:
+                                                  TextCapitalization.sentences,
                                               decoration: InputDecoration(
                                                 hintText: "Enter title... ",
-                                                hintStyle: Theme.of(context).textTheme.titleMedium?.copyWith(
-                                                    color: Theme.of(context).colorScheme.background,
-                                                    fontWeight: medium),
-                                                border: const UnderlineInputBorder(borderSide: BorderSide.none),
+                                                hintStyle: Theme.of(context)
+                                                    .textTheme
+                                                    .titleMedium
+                                                    ?.copyWith(
+                                                        color: Theme.of(context)
+                                                            .colorScheme
+                                                            .surface,
+                                                        fontWeight: medium),
+                                                border:
+                                                    const UnderlineInputBorder(
+                                                        borderSide:
+                                                            BorderSide.none),
                                               ),
                                             ),
                                           ),
@@ -452,9 +516,13 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                                             likeBuilder: (bool isLiked) {
                                               favourite = isLiked;
                                               return Icon(
-                                                isLiked ? Icons.star : Icons.star_border,
-                                                color:
-                                                    isLiked ? Theme.of(context).indicatorColor : AppColors.cFadedBlue,
+                                                isLiked
+                                                    ? Icons.star
+                                                    : Icons.star_border,
+                                                color: isLiked
+                                                    ? Theme.of(context)
+                                                        .indicatorColor
+                                                    : AppColors.cFadedBlue,
                                                 size: 36,
                                               );
                                             },
@@ -463,21 +531,32 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                                       ),
                                     ),
                                     Padding(
-                                      padding: const EdgeInsets.symmetric(horizontal: 16),
+                                      padding: const EdgeInsets.symmetric(
+                                          horizontal: 16),
                                       child: TextField(
                                         controller: _textController,
                                         keyboardType: TextInputType.multiline,
-                                        textCapitalization: TextCapitalization.sentences,
+                                        textCapitalization:
+                                            TextCapitalization.sentences,
                                         style: Theme.of(context)
                                             .textTheme
                                             .labelLarge
-                                            ?.copyWith(fontWeight: regular, height: 1.4),
+                                            ?.copyWith(
+                                                fontWeight: regular,
+                                                height: 1.4),
                                         maxLines: null,
                                         decoration: InputDecoration(
                                           hintText: "Enter new note... ",
-                                          hintStyle: Theme.of(context).textTheme.labelLarge?.copyWith(
-                                              color: Theme.of(context).colorScheme.background, fontWeight: regular),
-                                          border: const UnderlineInputBorder(borderSide: BorderSide.none),
+                                          hintStyle: Theme.of(context)
+                                              .textTheme
+                                              .labelLarge
+                                              ?.copyWith(
+                                                  color: Theme.of(context)
+                                                      .colorScheme
+                                                      .surface,
+                                                  fontWeight: regular),
+                                          border: const UnderlineInputBorder(
+                                              borderSide: BorderSide.none),
                                         ),
                                       ),
                                     ),
@@ -493,54 +572,92 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                                   child: ValueListenableBuilder(
                                     valueListenable: _imageFile,
                                     builder: (context, imageFile, _) {
-                                      return _imageFile.value == null || _imageFile.value?.path == null
+                                      return _imageFile.value == null ||
+                                              _imageFile.value?.path == null
                                           ? ValueListenableBuilder(
                                               valueListenable: _imageUrl,
                                               builder: (context, imageUrl, _) {
                                                 return _imageUrl.value != ""
                                                     ? Padding(
-                                                        padding: const EdgeInsets.fromLTRB(16, 16, 8, 16),
+                                                        padding:
+                                                            const EdgeInsets
+                                                                .fromLTRB(
+                                                                16, 16, 8, 16),
                                                         child: ClipRRect(
-                                                          borderRadius: BorderRadius.circular(15),
+                                                          borderRadius:
+                                                              BorderRadius
+                                                                  .circular(15),
                                                           child: Column(
                                                             children: [
                                                               InkWell(
                                                                 onTap: () {
                                                                   Utilities.openNamedActivity(
-                                                                      context, Routes.enlargeImage,
-                                                                      arguments: ImageArgs(imageUrl: _imageUrl.value));
+                                                                      context,
+                                                                      Routes
+                                                                          .enlargeImage,
+                                                                      arguments:
+                                                                          ImageArgs(
+                                                                              imageUrl: _imageUrl.value));
                                                                 },
-                                                                child: Container(
+                                                                child:
+                                                                    Container(
                                                                   height: 140,
-                                                                  width: double.maxFinite,
-                                                                  color: AppColors.cWhite,
-                                                                  child: CachedNetworkImage(
-                                                                    imageUrl: _imageUrl.value,
-                                                                    fit: BoxFit.cover,
-                                                                    memCacheHeight: 450,
-                                                                    errorWidget: (context, url, error) => Icon(
-                                                                      Icons.error,
-                                                                      color: Theme.of(context).colorScheme.background,
+                                                                  width: double
+                                                                      .maxFinite,
+                                                                  color: AppColors
+                                                                      .cWhite,
+                                                                  child:
+                                                                      CachedNetworkImage(
+                                                                    imageUrl:
+                                                                        _imageUrl
+                                                                            .value,
+                                                                    fit: BoxFit
+                                                                        .cover,
+                                                                    memCacheHeight:
+                                                                        450,
+                                                                    errorWidget: (context,
+                                                                            url,
+                                                                            error) =>
+                                                                        Icon(
+                                                                      Icons
+                                                                          .error,
+                                                                      color: Theme.of(
+                                                                              context)
+                                                                          .colorScheme
+                                                                          .surface,
                                                                     ),
                                                                   ),
                                                                 ),
                                                               ),
                                                               InkWell(
                                                                 onTap: () {
-                                                                  _imageFile.value = null;
-                                                                  if (_imageUrl.value != "") {
-                                                                    _notesService.deleteFile(_imageUrl.value);
-                                                                    _imageUrl.value = "";
+                                                                  _imageFile
+                                                                          .value =
+                                                                      null;
+                                                                  if (_imageUrl
+                                                                          .value !=
+                                                                      "") {
+                                                                    _notesService
+                                                                        .deleteFile(
+                                                                            _imageUrl.value);
+                                                                    _imageUrl
+                                                                        .value = "";
                                                                   }
                                                                 },
-                                                                child: Container(
+                                                                child:
+                                                                    Container(
                                                                   height: 42,
-                                                                  width: double.maxFinite,
-                                                                  color: AppColors.cRedAccent,
-                                                                  child: const Center(
+                                                                  width: double
+                                                                      .maxFinite,
+                                                                  color: AppColors
+                                                                      .cRedAccent,
+                                                                  child:
+                                                                      const Center(
                                                                     child: Icon(
-                                                                      Icons.close,
-                                                                      color: AppColors.cWhite,
+                                                                      Icons
+                                                                          .close,
+                                                                      color: AppColors
+                                                                          .cWhite,
                                                                     ),
                                                                   ),
                                                                 ),
@@ -553,42 +670,55 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                                               },
                                             )
                                           : Padding(
-                                              padding: const EdgeInsets.fromLTRB(16, 16, 8, 16),
+                                              padding:
+                                                  const EdgeInsets.fromLTRB(
+                                                      16, 16, 8, 16),
                                               child: ClipRRect(
-                                                borderRadius: BorderRadius.circular(15),
+                                                borderRadius:
+                                                    BorderRadius.circular(15),
                                                 child: Column(
                                                   children: [
                                                     Container(
                                                       height: 140,
                                                       width: double.maxFinite,
                                                       color: AppColors.cWhite,
-                                                      child: Image.file(_imageFile.value!,
-                                                          fit: BoxFit.cover, cacheHeight: 500),
+                                                      child: Image.file(
+                                                          _imageFile.value!,
+                                                          fit: BoxFit.cover,
+                                                          cacheHeight: 500),
                                                     ),
                                                     ValueListenableBuilder(
                                                       valueListenable: _task,
-                                                      builder: (context, task, _) {
+                                                      builder:
+                                                          (context, task, _) {
                                                         return task != null
-                                                            ? buildUploadStatus(_task.value!)
+                                                            ? buildUploadStatus(
+                                                                _task.value!)
                                                             : const SizedBox();
                                                       },
                                                     ),
                                                     InkWell(
                                                       onTap: () {
                                                         _imageFile.value = null;
-                                                        if (_imageUrl.value != "") {
-                                                          _notesService.deleteFile(_imageUrl.value);
+                                                        if (_imageUrl.value !=
+                                                            "") {
+                                                          _notesService
+                                                              .deleteFile(
+                                                                  _imageUrl
+                                                                      .value);
                                                           _imageUrl.value = "";
                                                         }
                                                       },
                                                       child: Container(
                                                         height: 42,
                                                         width: 200,
-                                                        color: AppColors.cRedAccent,
+                                                        color: AppColors
+                                                            .cRedAccent,
                                                         child: const Center(
                                                           child: Icon(
                                                             Icons.close,
-                                                            color: AppColors.cWhite,
+                                                            color: AppColors
+                                                                .cWhite,
                                                           ),
                                                         ),
                                                       ),
@@ -604,50 +734,70 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                                   child: ValueListenableBuilder(
                                     valueListenable: _file,
                                     builder: (context, fileUrl, _) {
-                                      return _file.value == null || _file.value?.path == null
+                                      return _file.value == null ||
+                                              _file.value?.path == null
                                           ? ValueListenableBuilder(
                                               valueListenable: _fileUrl,
                                               builder: (context, fileUrl, _) {
                                                 return _fileUrl.value != ""
                                                     ? Padding(
-                                                        padding: const EdgeInsets.fromLTRB(8, 16, 16, 16),
+                                                        padding:
+                                                            const EdgeInsets
+                                                                .fromLTRB(
+                                                                8, 16, 16, 16),
                                                         child: ClipRRect(
-                                                          borderRadius: BorderRadius.circular(15),
+                                                          borderRadius:
+                                                              BorderRadius
+                                                                  .circular(15),
                                                           child: Column(
                                                             children: [
                                                               InkWell(
                                                                 onTap: () {
-                                                                  Utilities.openNamedActivity(context, Routes.pdfView,
+                                                                  Utilities.openNamedActivity(
+                                                                      context,
+                                                                      Routes
+                                                                          .pdfView,
                                                                       arguments: Args(
-                                                                          fileUrl: _fileUrl.value,
-                                                                          fileName: widget.note!.fileName ?? ""));
+                                                                          fileUrl: _fileUrl
+                                                                              .value,
+                                                                          fileName:
+                                                                              widget.note!.fileName ?? ""));
                                                                 },
-                                                                child: Container(
+                                                                child:
+                                                                    Container(
                                                                   height: 140,
-                                                                  width: double.maxFinite,
-                                                                  color: AppColors.cWhite,
+                                                                  width: double
+                                                                      .maxFinite,
+                                                                  color: AppColors
+                                                                      .cWhite,
                                                                   child: Center(
-                                                                    child: Column(
-                                                                      mainAxisAlignment: MainAxisAlignment.center,
+                                                                    child:
+                                                                        Column(
+                                                                      mainAxisAlignment:
+                                                                          MainAxisAlignment
+                                                                              .center,
                                                                       children: [
                                                                         Icon(
-                                                                            path.extension(widget.note!.fileName!) ==
-                                                                                    ".pdf"
+                                                                            path.extension(widget.note!.fileName!) == ".pdf"
                                                                                 ? Icons.picture_as_pdf
-                                                                                : path.extension(
-                                                                                            widget.note!.fileName!) ==
-                                                                                        ".pptx"
+                                                                                : path.extension(widget.note!.fileName!) == ".pptx"
                                                                                     ? Icons.present_to_all
                                                                                     : Icons.file_copy,
                                                                             size: 48,
                                                                             color: AppColors.cDarkBlue),
-                                                                        const SizedBox(height: 10),
+                                                                        const SizedBox(
+                                                                            height:
+                                                                                10),
                                                                         Text(
                                                                           "${widget.note?.fileName}",
-                                                                          textAlign: TextAlign.center,
-                                                                          maxLines: 2,
-                                                                          style: CustomTextStyle.smallText,
-                                                                          overflow: TextOverflow.ellipsis,
+                                                                          textAlign:
+                                                                              TextAlign.center,
+                                                                          maxLines:
+                                                                              2,
+                                                                          style:
+                                                                              CustomTextStyle.smallText,
+                                                                          overflow:
+                                                                              TextOverflow.ellipsis,
                                                                         ),
                                                                       ],
                                                                     ),
@@ -656,21 +806,33 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                                                               ),
                                                               InkWell(
                                                                 onTap: () {
-                                                                  _file.value = null;
-                                                                  if (_fileUrl.value != "") {
-                                                                    _notesService.deleteFile(_fileUrl.value);
-                                                                    _fileUrl.value = "";
-                                                                    _fileName.value = "";
+                                                                  _file.value =
+                                                                      null;
+                                                                  if (_fileUrl
+                                                                          .value !=
+                                                                      "") {
+                                                                    _notesService
+                                                                        .deleteFile(
+                                                                            _fileUrl.value);
+                                                                    _fileUrl.value =
+                                                                        "";
+                                                                    _fileName
+                                                                        .value = "";
                                                                   }
                                                                 },
-                                                                child: Container(
+                                                                child:
+                                                                    Container(
                                                                   height: 42,
                                                                   width: 200,
-                                                                  color: AppColors.cRedAccent,
-                                                                  child: const Center(
+                                                                  color: AppColors
+                                                                      .cRedAccent,
+                                                                  child:
+                                                                      const Center(
                                                                     child: Icon(
-                                                                      Icons.close,
-                                                                      color: AppColors.cWhite,
+                                                                      Icons
+                                                                          .close,
+                                                                      color: AppColors
+                                                                          .cWhite,
                                                                     ),
                                                                   ),
                                                                 ),
@@ -683,9 +845,12 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                                               },
                                             )
                                           : Padding(
-                                              padding: const EdgeInsets.fromLTRB(8, 16, 16, 16),
+                                              padding:
+                                                  const EdgeInsets.fromLTRB(
+                                                      8, 16, 16, 16),
                                               child: ClipRRect(
-                                                borderRadius: BorderRadius.circular(15),
+                                                borderRadius:
+                                                    BorderRadius.circular(15),
                                                 child: Column(
                                                   children: [
                                                     Container(
@@ -693,40 +858,68 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                                                       width: double.maxFinite,
                                                       color: AppColors.cWhite,
                                                       child: Column(
-                                                        mainAxisAlignment: MainAxisAlignment.center,
+                                                        mainAxisAlignment:
+                                                            MainAxisAlignment
+                                                                .center,
                                                         children: [
                                                           Icon(
-                                                              path.extension(_file.value!.path) == ".pdf"
-                                                                  ? Icons.picture_as_pdf
-                                                                  : path.extension(_file.value!.path) == ".pptx"
-                                                                      ? Icons.present_to_all
-                                                                      : Icons.file_copy,
+                                                              path.extension(_file
+                                                                          .value!
+                                                                          .path) ==
+                                                                      ".pdf"
+                                                                  ? Icons
+                                                                      .picture_as_pdf
+                                                                  : path.extension(_file
+                                                                              .value!
+                                                                              .path) ==
+                                                                          ".pptx"
+                                                                      ? Icons
+                                                                          .present_to_all
+                                                                      : Icons
+                                                                          .file_copy,
                                                               size: 48,
-                                                              color: AppColors.cDarkBlue),
-                                                          const SizedBox(height: 10),
+                                                              color: AppColors
+                                                                  .cDarkBlue),
+                                                          const SizedBox(
+                                                              height: 10),
                                                           Text(
-                                                            _file.value!.path.split("/").last,
-                                                            textAlign: TextAlign.center,
+                                                            _file.value!.path
+                                                                .split("/")
+                                                                .last,
+                                                            textAlign: TextAlign
+                                                                .center,
                                                             maxLines: 2,
-                                                            style: CustomTextStyle.smallText,
-                                                            overflow: TextOverflow.ellipsis,
+                                                            style:
+                                                                CustomTextStyle
+                                                                    .smallText,
+                                                            overflow:
+                                                                TextOverflow
+                                                                    .ellipsis,
                                                           )
                                                         ],
                                                       ),
                                                     ),
                                                     ValueListenableBuilder(
-                                                      valueListenable: _taskFile,
-                                                      builder: (context, task, _) {
+                                                      valueListenable:
+                                                          _taskFile,
+                                                      builder:
+                                                          (context, task, _) {
                                                         return task != null
-                                                            ? buildFileUploadStatus(_taskFile.value!)
+                                                            ? buildFileUploadStatus(
+                                                                _taskFile
+                                                                    .value!)
                                                             : const SizedBox();
                                                       },
                                                     ),
                                                     InkWell(
                                                       onTap: () {
                                                         _file.value = null;
-                                                        if (_fileUrl.value != "") {
-                                                          _notesService.deleteFile(_fileUrl.value);
+                                                        if (_fileUrl.value !=
+                                                            "") {
+                                                          _notesService
+                                                              .deleteFile(
+                                                                  _fileUrl
+                                                                      .value);
                                                           _fileUrl.value = "";
                                                           _fileName.value = "";
                                                         }
@@ -734,9 +927,13 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
                                                       child: Container(
                                                         height: 42,
                                                         width: 200,
-                                                        color: AppColors.cRedAccent,
+                                                        color: AppColors
+                                                            .cRedAccent,
                                                         child: const Center(
-                                                          child: Icon(Icons.close, color: AppColors.cWhite),
+                                                          child: Icon(
+                                                              Icons.close,
+                                                              color: AppColors
+                                                                  .cWhite),
                                                         ),
                                                       ),
                                                     ),
@@ -794,7 +991,8 @@ class CreateUpdateNoteViewState extends State<CreateUpdateNoteView> {
   Future uploadFile() async {
     if (_file.value == null) return;
     final fileName = basename(_file.value!.path);
-    _taskFile.value = _notesService.uploadFile("$fileName${DateTime.now()}", _file.value!);
+    _taskFile.value =
+        _notesService.uploadFile("$fileName${DateTime.now()}", _file.value!);
 
     if (_taskFile.value == null) return;
     final snapshot = await _taskFile.value!.whenComplete(() {});
