@@ -1,3 +1,4 @@
 extension Filter<T> on Stream<List<T>> {
-  Stream<List<T>> filter(bool Function(T) where) => map((items) => items.where(where).toList());
+  Stream<List<T>> filter(bool Function(T) where) =>
+      map((items) => items.where(where).toList());
 }
