@@ -1,7 +1,9 @@
 import 'package:flutter/rendering.dart';
 
-const String privacyPolicyUrl = "https://www.freeprivacypolicy.com/live/ec30db4f-ca00-48fa-9352-cb635aa738fa";
-const String termsUrl = "https://www.termsfeed.com/live/2afbfedc-d181-4938-a22b-c56d8afe0307";
+const String privacyPolicyUrl =
+    "https://www.freeprivacypolicy.com/live/ec30db4f-ca00-48fa-9352-cb635aa738fa";
+const String termsUrl =
+    "https://www.termsfeed.com/live/2afbfedc-d181-4938-a22b-c56d8afe0307";
 const String appName = "Notester";
 
 const FontWeight light = FontWeight.w300;
