@@ -1,19 +1,23 @@
-import 'package:notester/provider/dark_theme_provider.dart';
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
 
 class LogoWidget extends StatelessWidget {
   final double? height;
   final double? width;
-  const LogoWidget({Key? key, this.height, this.width}) : super(key: key);
+  final BoxFit? boxFit;
+  const LogoWidget({Key? key, this.height, this.width, this.boxFit})
+      : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    final themeProvider = context.watch<DarkThemeProvider>();
+    // final themeProvider = context.watch<DarkThemeProvider>();
     return Image.asset(
-      themeProvider.darkTheme ? "assets/appLogo.png" : "assets/appLogoLight.png",
+      "assets/notesterLogoTransparent.png",
+      // themeProvider.darkTheme
+      //     ? "assets/appLogo.png"
+      //     : "assets/appLogoLight.png",
       height: height,
       width: width,
+      fit: boxFit,
     );
   }
 }
