@@ -8,7 +8,14 @@ class SafeAreaHeader extends StatelessWidget {
   final bool? left;
   final bool? right;
 
-  const SafeAreaHeader({Key? key, required this.child, this.color, this.top, this.bottom, this.left, this.right})
+  const SafeAreaHeader(
+      {Key? key,
+      required this.child,
+      this.color,
+      this.top,
+      this.bottom,
+      this.left,
+      this.right})
       : super(key: key);
   @override
   Widget build(BuildContext context) {
